refactor: common `StickyBar` component
